# mat: add fast paths to matrix sum function