@kortschak 2019-04-03
<table class='benchstat'>
<tr><th>name</th><th>old time/op</th><th>new time/op</th><th>delta</th>
<tr><td>DenseSum1000-8</td><td>1.25ms ± 2%</td><td>1.24ms ± 3%</td><td>~   </td><td>(p=0.356 n=10+9)</td>
<tr><td>SymSum1000-8</td><td>5.24ms ± 9%</td><td>0.60ms ± 2%</td><td>-88.61%</td><td>(p=0.000 n=10+10)</td>
<tr><td>TriSum1000-8</td><td>4.28ms ± 6%</td><td>0.61ms ± 1%</td><td>-85.73%</td><td>(p=0.000 n=10+10)</td>
<tr><td>VectorSum100000-8</td><td>412µs ± 2%</td><td>119µs ± 1%</td><td>-71.13%</td><td>(p=0.000 n=9+10)</td>
</table>