@sbinet 2019-04-03
ah! didn't know `benchstat` could output HTML :)

nice speedups!
congrats @jonreiter (and thanks again!)